Lazy load generatedResourcesFoldername
When using this from a maven plugin, the getting the ClassPathResource
path fails, causing an error to be logged, even if the path is going to
be set explicitly.
By moving the initialization of the path to the getter, trying to get
the path automatically can be omitted if it is set explicitly, avoiding
the error message that doesn't make any sense in this case.